cheng update project

+++ admin/src/main/kotlin/acheng1314/cn/cache/AdminUserCacheExtend.kt
@@ -18,15 +18,15 @@ class AdminUserCacheExtend {
      * 获取用户权限，从缓存中获取
      */
     suspend fun getMenusFromCache(admin: AdminUserDTO): MutableList<AdminMenuDTO> {
-        val menuIds: List<Int>?
+        val menuIds: List<Long>?
         val menus = mutableListOf<AdminMenuDTO>()
         val duty = admin.duty
         menuIds = try {
-            JSONUtil.toArray(duty, Int::class.java)
+            JSONUtil.toArray(duty, Long::class.java)
         } catch (e: Exception) {
             val rule = siteConfigService.getDefualtRuleByName(duty)
             logger.info("用户${admin.loginName}没有自定义权限，默认权限为：$duty")
-            JSONUtil.toArray(rule, Int::class.java)
+            JSONUtil.toArray(rule, Long::class.java)
         }
         if (menuIds.isNullOrEmpty()) throw Exception("该用户暂无权限配置！")
         if (Cache.adminMenuIdCache.count() == 0) adminMenuService.notifyMenu()
